refactor(web): rename service for right convention
